microstructure-analysis-flask/requirements.txt
Flask==2.0.1
Pillow==9.0.0
PyYAML==5.4.1
Flask-SQLAlchemy==2.5.1
pytest==6.2.5
python-dotenv==0.19.0
Werkzeug==2.0.1 